Refactor conversation service models
